refactor:Converter 생성하여 역할 분리
